Added PSR-4 compliance & refactor to use SDK 1.0.0 + API v2
