+++ composer.json
@@ -22,7 +22,8 @@
     "require": {
         "php":                          ">=5.3.3",
         "firebase/php-jwt":             "dev-master",
-        "adoy/oauth2":                  "dev-master"
+        "adoy/oauth2":                  "dev-master",
+        "auth0/auth0-php":              "~1.0"
     },
 
     "require-dev": {
@@ -32,10 +33,8 @@
     },
 
     "autoload": {
-        "psr-0": {
-            "Auth0\\JWTAuthBundle\\": ""
+        "psr-4": {
+            "Auth0\\JWTAuthBundle\\": "src/"
         }
-    },
-
-    "target-dir": "Auth0/JWTAuthBundle"
+    }
 }
