+++ example/src/AppBundle/Resources/public/js/auth0-variables.js
